+++ test/td_dq_web/controllers/implementation_controller_test.exs
@@ -843,7 +843,7 @@ defmodule TdDqWeb.ImplementationControllerTest do
       assert {:ok, %{implementation: %{id: ^id}}} =
                Implementations.update_implementation(
                  implementation,
-                 %{status: :published},
+                 %{status: :published, minimum: implementation.minimum - 1},
                  claims
                )
 
@@ -889,7 +889,7 @@ defmodule TdDqWeb.ImplementationControllerTest do
       assert {:ok, %{implementation: %{id: ^id, version: 1, status: :published}}} =
                Implementations.update_implementation(
                  implementation,
-                 %{status: :published},
+                 %{status: :published, minimum: implementation.minimum - 1},
                  claims
                )
 
@@ -905,6 +905,42 @@ defmodule TdDqWeb.ImplementationControllerTest do
 
       assert %{implementation_ref: ^imp_ref, version: 2, status: :draft} =
                Implementations.get_implementation!(new_id)
+    end
+
+    @tag authentication: [role: "admin"]
+    test "return 200 but does not create a new draft when editing published implementation with same information",
+         %{
+           conn: conn,
+           swagger_schema: schema,
+           claims: claims
+         } do
+      rule_implementation_attr = string_params_for(:implementation)
+
+      assert %{"data" => %{"id" => id}} =
+               conn
+               |> post(Routes.implementation_path(conn, :create),
+                 rule_implementation: rule_implementation_attr
+               )
+               |> json_response(:created)
+
+      implementation = Implementations.get_implementation(id)
+
+      Implementations.update_implementation(
+        implementation,
+        %{status: :published},
+        claims
+      )
+
+      assert %{
+               "data" => %{"id" => ^id, "status" => "published"},
+               "message" => "implementation_unchanged"
+             } =
+               conn
+               |> put(Routes.implementation_path(conn, :update, implementation),
+                 rule_implementation: rule_implementation_attr
+               )
+               |> validate_resp_schema(schema, "ImplementationResponse")
+               |> json_response(:ok)
     end
 
     @tag authentication: [role: "admin"]